+++ packages/tailwindcss-language-server/tests/diagnostics/diagnostics.test.js
@@ -95,6 +95,7 @@ withFixture('v4/basic', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 36 } },
         severity: 1,
         message: "'--color-red-901' does not exist in your theme. Did you mean '--color-red-900'?",
@@ -109,6 +110,7 @@ withFixture('v4/basic', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 37 } },
         severity: 1,
         message: "'--colors-red-901' does not exist in your theme. Did you mean '--color-red-900'?",
@@ -123,6 +125,7 @@ withFixture('v4/basic', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 42 } },
         severity: 1,
         message: "'--font-obliqueness-90' does not exist in your theme.",
@@ -143,6 +146,7 @@ withFixture('v4/basic', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 35 } },
         severity: 1,
         message: "'colors.red.901' does not exist in your theme config.",
@@ -194,6 +198,7 @@ withFixture('v4/with-prefix', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 36 } },
         severity: 1,
         message: "'--color-red-901' does not exist in your theme. Did you mean '--color-red-900'?",
@@ -208,6 +213,7 @@ withFixture('v4/with-prefix', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 37 } },
         severity: 1,
         message: "'--colors-red-901' does not exist in your theme. Did you mean '--color-red-900'?",
@@ -222,6 +228,7 @@ withFixture('v4/with-prefix', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 42 } },
         severity: 1,
         message: "'--font-obliqueness-90' does not exist in your theme.",
@@ -242,6 +249,7 @@ withFixture('v4/with-prefix', (c) => {
     expected: [
       {
         code: 'invalidConfigPath',
+        source: 'tailwindcss',
         range: { start: { line: 0, character: 21 }, end: { line: 0, character: 35 } },
         severity: 1,
         message: "'colors.red.901' does not exist in your theme config.",
@@ -280,6 +288,7 @@ withFixture('v4/basic', (c) => {
     expected: [
       {
         code: 'cssConflict',
+        source: 'tailwindcss',
         message: "'max-w-4xl' applies the same CSS properties as 'max-w-6xl'.",
         className: {
           className: 'max-w-4xl',
@@ -298,6 +307,7 @@ withFixture('v4/basic', (c) => {
       },
       {
         code: 'cssConflict',
+        source: 'tailwindcss',
         message: "'max-w-6xl' applies the same CSS properties as 'max-w-4xl'.",
         className: {
           className: 'max-w-6xl',
@@ -329,6 +339,7 @@ withFixture('v4/basic', (c) => {
     expected: [
       {
         code: 'invalidTailwindDirective',
+        source: 'tailwindcss',
         message:
           "'@tailwind base' is no longer available in v4. Use '@import \"tailwindcss/preflight\"' instead.",
         suggestions: [],
@@ -340,6 +351,7 @@ withFixture('v4/basic', (c) => {
       },
       {
         code: 'invalidTailwindDirective',
+        source: 'tailwindcss',
         message:
           "'@tailwind preflight' is no longer available in v4. Use '@import \"tailwindcss/preflight\"' instead.",
         suggestions: [],
@@ -351,6 +363,7 @@ withFixture('v4/basic', (c) => {
       },
       {
         code: 'invalidTailwindDirective',
+        source: 'tailwindcss',
         message:
           "'@tailwind components' is no longer available in v4. Use '@tailwind utilities' instead.",
         suggestions: ['utilities'],
@@ -362,6 +375,7 @@ withFixture('v4/basic', (c) => {
       },
       {
         code: 'invalidTailwindDirective',
+        source: 'tailwindcss',
         message:
           "'@tailwind screens' is no longer available in v4. Use '@tailwind utilities' instead.",
         suggestions: ['utilities'],
@@ -373,6 +387,7 @@ withFixture('v4/basic', (c) => {
       },
       {
         code: 'invalidTailwindDirective',
+        source: 'tailwindcss',
         message:
           "'@tailwind variants' is no longer available in v4. Use '@tailwind utilities' instead.",
         suggestions: ['utilities'],
@@ -406,6 +421,7 @@ defineTest({
     expect(diagnostics).toEqual([
       {
         code: 'usedBlocklistedClass',
+        source: 'tailwindcss',
         message: 'The class "flex" will not be generated as it has been blocklisted',
         range: {
           start: { line: 0, character: 12 },
@@ -415,6 +431,7 @@ defineTest({
       },
       {
         code: 'usedBlocklistedClass',
+        source: 'tailwindcss',
         message: 'The class "hover:flex" will not be generated as it has been blocklisted',
         range: {
           start: { line: 0, character: 27 },
@@ -463,6 +480,7 @@ defineTest({
     expect(diagnostics).toEqual([
       {
         code: 'suggestCanonicalClasses',
+        source: 'tailwindcss',
         message: 'The class `[@media_print]:flex` can be written as `print:flex`',
         range: {
           start: { line: 0, character: 12 },
@@ -473,6 +491,7 @@ defineTest({
       },
       {
         code: 'suggestCanonicalClasses',
+        source: 'tailwindcss',
         message: 'The class `[color:red]/50` can be written as `text-[red]/50`',
         range: {
           start: { line: 0, character: 32 },
@@ -483,6 +502,7 @@ defineTest({
       },
       {
         code: 'suggestCanonicalClasses',
+        source: 'tailwindcss',
         message: 'The class `mt-[16px]` can be written as `mt-4`',
         range: {
           start: { line: 0, character: 47 },